server/requirements.txt
# Core Django packages
Django>=5.0.2,<5.1
djangorestframework>=3.14.0,<4.0
django-cors-headers>=4.3.0,<5.0

# Environment configuration
python-dotenv>=1.0.0,<2.0

# Firebase
firebase-admin>=6.2.0,<7.0

# AI/ML APIs
anthropic>=0.25.0,<1.0
openai>=1.12.0,<2.0
google-genai>=0.5.0,<1.0
psycopg2
# HTTP requests
requests>=2.31.0,<3.0

# Production server (optional)
gunicorn>=21.2.0,<22.0
